[3D][Fixed]: Do not fail resistor coloring when model is not wrl
The resistor coloring code only works with WRL files, but it did not check
that the 3D model is actually WRL. So when a footprint used e.g. a STEP
file (such as the default KiCad resistor footprints), it would read the
STEP file and write it unmodified into an WRL file and use that.
Rendering with the blender output would then have the resistor missing.

This commit changes the code to:
 - Use the WRL version of the model if it exists, resulting in properly
   colored resistors in the rendering.
 - Log a warning if no WRL version is available and use the original
   (uncolored) STEP file.

This fixes #856