Support updating text variables in .kicad_pcb file explicitly
Normally, the text variables in the .kicad_pcb file are cleared when
requested, or when changing them with the set_text_variables preflight.
This causes KiCad to load the variables from the project file when
opening the PCB, but this does not work when the .kicad_pcb file is used
in isolation (e.g. when submitting to a PCB manufacturer).

This commit adds support for a new update_pcb_text_cache option, which
causes kibot to explicitly write the (updated / current) variables to
the .kicad_pcb file, making it self-contained.

This fixes #860